requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.0
google-genai==1.9.0
google-cloud-storage==2.16.0
google-cloud-monitoring==2.20.0
google-auth==2.29.0
gitpython==3.1.43
pip-audit==2.7.3
PyYAML==6.0.1
python-multipart==0.0.22
soul-agent==0.1.1
